requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
pandas>=1.5.0
openpyxl>=3.0.0
gunicorn>=20.1.0
whitenoise>=6.0.0
